Add code necessary for the snap install to succeed
We have to let the snapd service initialize and we have to start the
snapd socket before we can attempt to install anything via snap.
Additionally, the snap install command tends to fail the first time we
run it when testing via Molecule so we have to re-run the command
until it succeeds.  (Usually only a second run is necessary.)